voucher

+++ ms_contact_management/infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/bancolombia/newness/data/NewnessData.java
@@ -29,7 +29,7 @@ public class NewnessData {
     private String userCreation;
     private String responseCode;
     private String responseDescription;
-    private Integer voucher;
+    private Long voucher;
     private LocalDateTime dateFirstInscription;
     private LocalDateTime dateCreation;
 }
